Enable Nullability Information in IOperation.Type
This enables nullability information from rewritten BoundNodes
in IOperation, exposed via the NullableAnnotation property on
IOperation.Type. The bulk of the change is creating a simple helper
on BoundExpression to return the correctly annotated ITypeSymbol,
and using that in the factory instead of .Type.GetPublicSymbol. This
exposed another issue with a few of our tests though: in DEBUG mode,
when nullable is gobally disabled, we perform a full nullable analysis
and rewrite and then discard the results, to ensure that we don't
crash and to run our general verification code. However, there are
a few nodes that get cached by binders (such as LockOrUsingBinder).
These nodes do not get rewritten as their constituent elements did
not change, so when we set lazy properties like TopLevelNullability
info, the cache from the throwaway binding gets updated as well.
To remedy this and still get coverage, we pass along a flag to abort
before the actual rewrite in DEBUG scenarios.